introducing two CPT examples with readme; cleaned up repo readme; package version uptick
