+++ post-type/README.md
@@ -1,0 +1,6 @@
+### Custom Post Type Examples
+The following houses the directory where any CPT will live. Files will be called via `require_once` from `functions.php` file. To be used as examples and/or base for building any CPT for a project.
+
+#### Directory Content
+* Regular CPT
+* CPT with meta fields, meta field saving, custom column from any special meta created.
